finish project

+++ model/product.go
@@ -1,7 +1,7 @@
 package model
 
 type Product struct {
-	ID    string  `json:"id"`
-	NAME  string  `json:"name"`
-	PRICE float32 `json:"price"`
+	ID    string `json:"id"`
+	NAME  string `json:"name"`
+	PRICE int    `json:"price"`
 }
